requirements.txt

gradio==4.44.0
faiss-cpu>=1.12.0,<2
sentence-transformers==3.0.1
transformers>=4.41.0
pypdf==4.2.0
openai==1.46.0
numpy==1.26.4
rank_bm25==0.2.2
